api/requirements.txt
tensorflow==2.10.0
tensorflow-serving-api==2.10.0
fastapi
uvicorn[standard]
python-multipart
pillow
matplotlib
numpy
